refactors archive route to use the same function for both html and json requests #191
